fix Intersection of periods update
