+++ modules/Attendance/Requests/UpdateAttendanceConstraintRequest.php
@@ -8,9 +8,14 @@ use Illuminate\Foundation\Http\FormRequest;
 use Modules\Attendance\Models\AttendanceConstraint;
 use Modules\Attendance\DTO\UpdateAttendanceConstraintDTO;
 use Ramsey\Uuid\UuidInterface;
+use Modules\Attendance\Requests\Traits\HasConstraintConfigValidation;
+use InvalidArgumentException;
+use Modules\Attendance\DataClasses\MultiplePeriodsConfig;
+
 
 class UpdateAttendanceConstraintRequest extends FormRequest
 {
+     use HasConstraintConfigValidation;
     /**
      * Determine if the user is authorized to make this request.
      */
@@ -25,34 +30,41 @@ class UpdateAttendanceConstraintRequest extends FormRequest
     public function rules(): array
     {
         return [
-            'user_ids'         => ['nullable', 'array'],
+            // UIDs for relationships should be 'nullable' and 'array' if not always sent,
+            // but 'uuid' and 'exists' if sent. 'sometimes' means apply if present.
+            'user_ids'         => ['nullable', 'array', 'sometimes'],
             'user_ids.*'       => ['uuid', 'exists:users,id'],
-            'department_ids'   => ['nullable', 'array'],
+            'department_ids'   => ['nullable', 'array', 'sometimes'],
             'department_ids.*' => ['exists:management_hierarchies,id'],
-            'branch_ids' => ['nullable', 'array'],
+            'branch_ids' => ['nullable', 'array', 'sometimes'],
             'branch_ids.*' => ['exists:management_hierarchies,id'],
-            'branch_locations' => ['nullable', 'array'],
-            'branch_locations.*' => ['array'],
+
+            // branch_locations structure. Make top level 'sometimes|nullable|array'
+            'branch_locations' => ['sometimes', 'nullable', 'array'],
+            // Sub-rules for branch_locations.* apply IF 'branch_locations' is present AND an array.
+            // 'required_with:branch_locations.*' implies that if branch_locations array has elements,
+            // these fields are required for each element. This is generally correct.
             'branch_locations.*.name' => ['required_with:branch_locations.*', 'string', 'max:255'],
             'branch_locations.*.branch_id' => ['required_with:branch_locations.*', 'max:50'],
             'branch_locations.*.address' => ['required_with:branch_locations.*', 'string', 'max:500'],
             'branch_locations.*.latitude' => ['required_with:branch_locations.*', 'numeric', 'between:-90,90'],
             'branch_locations.*.longitude' => ['required_with:branch_locations.*', 'numeric', 'between:-180,180'],
             'branch_locations.*.radius' => ['required_with:branch_locations.*', 'integer', 'min:1', 'max:10000'],
+
             'constraint_type' => [
-                'nullable',
-                'required',
+                'sometimes', // يمكن أن يتم تحديثه أحيانًا
+                'required',  // إذا تم إرساله، فهو مطلوب (ليس null)
                 'string',
-                // 'in:' . implode(',', array_keys(AttendanceConstraint::getConstraintArrayTypes()))
+                'in:' . implode(',', array_keys(AttendanceConstraint::getConstraintArrayTypes()))
             ],
             'constraint_name' => 'sometimes|required|string|max:255',
             'constraint_config' => 'sometimes|required|array',
-            'is_active' => 'boolean',
-            'inherit_from_parent' => ['boolean'],
-            'priority' => 'integer|min:1|max:10',
-            'start_date' => 'nullable|date',
-            'end_date' => 'nullable|date|after:start_date',
-            'notes' => 'nullable|string|max:1000',
+            'is_active' => 'sometimes|boolean',
+            'inherit_from_parent' => ['sometimes', 'boolean'],
+            'priority' => 'sometimes|integer|min:1|max:10',
+            'start_date' => 'nullable|date|sometimes',
+            'end_date' => 'nullable|date|after:start_date|sometimes',
+            'notes' => 'nullable|string|max:1000|sometimes',
         ];
     }
 
@@ -63,39 +75,43 @@ class UpdateAttendanceConstraintRequest extends FormRequest
     {
         return [
             'constraint_type.in' => 'The selected constraint type is invalid.',
-            'constraint_config.required' => 'Constraint configuration is required.',
+            'constraint_config.required' => 'Constraint configuration is required when provided.', // Updated message
             'end_date.after' => 'End date must be after start date.',
             'priority.min' => 'Priority must be at least 1.',
             'priority.max' => 'Priority cannot exceed 10.',
+            'constraint_config.time_rules.weekly_schedule.required' => 'The weekly schedule is required for multiple periods configuration.',
+            'branch_locations.required_when_location_enabled' => 'Branch locations are required when location attendance type is enabled.',
+            'branch_locations.*.name.required_with' => 'Branch location name is required.',
+            'branch_locations.*.branch_id.required_with' => 'Branch location ID is required.',
+            'branch_locations.*.address.required_with' => 'Branch location address is required.',
+            'branch_locations.*.latitude.required_with' => 'Branch location latitude is required.',
+            'branch_locations.*.longitude.required_with' => 'Branch location longitude is required.',
+            'branch_locations.*.radius.required_with' => 'Branch location radius is required and must be a positive integer.',
         ];
     }
 
-    /**
+   /**
      * Configure the validator instance.
      */
-    // public function withValidator($validator): void
-    // {
-    //     $validator->after(function ($validator) {
-    //         if ($this->has('constraint_config')) {
-    //             $this->validateConstraintConfig($validator);
-    //         }
-    //     });
-    // }
+    public function withValidator($validator): void
+    {
+        $validator->after(function ($validator) {
+            // Only validate constraint_config details if it's explicitly present in the request
+            // and if constraint_type is also present (as config validation depends on type)
+            if ($this->has('constraint_config') && $this->has('constraint_type')) {
+                // Call the shared validation logic from the trait
+                $this->validateConstraintConfig($validator);
+            }
 
-    /**
-     * Validate constraint configuration based on type and name.
-     */
-    // protected function validateConstraintConfig($validator): void
-    // {
-    //     $type = $this->input('constraint_type');
-    //     $name = $this->input('constraint_name');
-    //     $config = $this->input('constraint_config', []);
+            // Additional cross-field validation for dates if both are present
+            if ($this->has('start_date') && $this->has('end_date')) {
+                if ($this->input('start_date') && $this->input('end_date') && $this->input('start_date') >= $this->input('end_date')) {
+                    $validator->errors()->add('end_date', 'End date must be after start date.');
+                }
+            }
+        });
+    }
 
-    //     // Use the same validation logic as CreateAttendanceConstraintRequest
-    //     $createRequest = new CreateAttendanceConstraintRequest();
-    //     $createRequest->merge($this->all());
-    //     $createRequest->validateConstraintConfig($validator);
-    // }
 
     /**
      * Create DTO from validated request data.
@@ -110,9 +126,9 @@ class UpdateAttendanceConstraintRequest extends FormRequest
             name: $validated['constraint_name'] ?? null,
             notes: $validated['notes'] ?? null,
             config: $validated['constraint_config'] ?? [],
-            user_ids: $validated['user_ids'] ?? [],
-            department_ids: $validated['department_ids'] ?? [],
-            branch_ids: $validated['branch_ids'] ?? [],
+            user_ids: $validated['user_ids'] ?? null,
+            department_ids: $validated['department_ids'] ?? null,
+            branch_ids: $validated['branch_ids'] ?? null,
             branch_locations: $validated['branch_locations'] ?? null,
             priority: $validated['priority'] ?? null,
             is_active: $validated['is_active'] ?? null,
